+++ Daedalus/Rotation/Apollo.cs
@@ -1,4 +1,4 @@
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using Dalamud.Game.ClientState.JobGauge;
 using Dalamud.Game.ClientState.Objects.SubKinds;
 using Dalamud.Game.ClientState.Party;
@@ -80,6 +80,7 @@ public sealed class Apollo : BaseHealerRotation<IApolloContext, IApolloModule>
         _trainingService = trainingService;
 
         _scheduler = new RotationScheduler(actionService, jobGauges, configuration, timelineService, errorMetrics);
+        _scheduler.TargetingService = targetingService; // enables the corpse hard-target swap for raises
 
         _statusHelper = new StatusHelper();
         _partyHelper = new PartyHelper(objectTable, partyList, hpPredictionService, configuration);